+++ Tracker/Extensions/UICollectionView+Extensions.swift
@@ -11,7 +11,14 @@ extension UICollectionView {
         let cellSpacing: CGFloat
         let paddingWidth: CGFloat
         
-        init(cellCount: CGFloat, leftInset: CGFloat, rightInset: CGFloat, topInset: CGFloat, bottomInset: CGFloat, height: CGFloat, cellSpacing: CGFloat) {
+        init(cellCount: CGFloat,
+             leftInset: CGFloat,
+             rightInset: CGFloat,
+             topInset: CGFloat,
+             bottomInset: CGFloat,
+             height: CGFloat,
+             cellSpacing: CGFloat
+        ) {
             self.cellCount = cellCount
             self.leftInset = leftInset
             self.rightInset = rightInset
