+++ Tracker/Services/Protocols/StatServiceProtocol.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+// MARK: - StatService + StatServiceProtocol
+
+protocol StatServiceProtocol {
+    func getStatistic() -> [StatModel]
+}
