+++ Tracker/Services/Protocols/UserAppSettingsStorageProtocol.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+// MARK: - UserAppSettingsStorage + UserAppSettingsStorageProtocol
+
+protocol UserAppSettingsStorageProtocol {
+    var isOnboardingVisited: Bool { get set }
+    func clean()
+}
